Routine

+++ jogl-samples/src/tests/gl320/Gl_320_buffer_update.java
@@ -233,7 +233,7 @@ public class Gl_320_buffer_update extends Test {
                     GL_UNIFORM_BUFFER, 0, 16 * GLBuffers.SIZEOF_FLOAT,
                     GL_MAP_WRITE_BIT | GL_MAP_INVALIDATE_BUFFER_BIT);
             projection = FloatUtil.makePerspective(projection, 0, true, (float) Math.PI * 0.25f,
-                    glWindow.getWidth() / glWindow.getHeight(), 0.1f, 100.0f);
+                    (float) glWindow.getWidth() / glWindow.getHeight(), 0.1f, 100.0f);
             FloatUtil.makeIdentity(model);
             // update view matrix
             view();
@@ -272,9 +272,9 @@ public class Gl_320_buffer_update extends Test {
 
         GL3 gl3 = (GL3) gl;
 
-        gl3.glDeleteBuffers(Buffer.max.ordinal(),  bufferName, 0);
+        gl3.glDeleteBuffers(Buffer.max.ordinal(), bufferName, 0);
         gl3.glDeleteProgram(programName);
-        gl3.glDeleteVertexArrays(1,  vertexArrayName, 0);
+        gl3.glDeleteVertexArrays(1, vertexArrayName, 0);
 
         return true;
     }
